requirements.txt
aiosignal==1.2.0
alembic==1.7.7
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
async-generator==1.10
attrs==22.2.0
backcall==0.2.0
bayesian-optimization==1.4.0
bleach==4.1.0
blis==0.7.10
Boruta==0.3
catalogue==1.0.2
certifi==2021.5.30
cffi==1.15.1
charset-normalizer==2.0.12
click==8.0.4
cloudpickle==2.2.1
colorama==0.4.4
colorlover==0.3.0
cufflinks==0.17.3
cycler==0.11.0
cymem==2.0.7
databricks-cli==0.17.7
dataclasses==0.8
decorator==4.4.2
defusedxml==0.7.1
distlib==0.3.7
docker==5.0.3
entrypoints==0.4
filelock==3.4.1
Flask==2.0.3
frozenlist==1.2.0
funcy==2.0
future==0.18.3
gensim==3.8.3
gitdb==4.0.9
GitPython==3.1.18
greenlet==2.0.2
grpcio==1.48.2
gunicorn==21.2.0
htmlmin==0.1.12
idna==3.4
ImageHash==4.3.1
imbalanced-learn==0.7.0
importlib-metadata==4.8.3
importlib-resources==5.4.0
ipykernel==5.5.6
ipython==7.16.3
ipython-genutils==0.2.0
ipywidgets==7.7.5
itsdangerous==2.0.1
jedi==0.17.2
Jinja2==3.0.3
joblib==1.0.1
jsonschema==3.2.0
jupyter-client==7.1.2
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.1.4
kiwisolver==1.3.1
kmodes==0.12.2
lightgbm==3.3.5
llvmlite==0.36.0
Mako==1.1.6
MarkupSafe==2.0.1
matplotlib==3.3.4
missingno==0.5.2
mistune==0.8.4
mlflow==1.23.1
mlxtend==0.19.0
modin==0.8.3.post4
msgpack==1.0.5
multimethod==1.5
murmurhash==1.0.9
mypy-extensions==1.0.0
nbclient==0.5.9
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.6
networkx==2.5.1
nltk==3.6.7
notebook==6.4.10
numba==0.53.1
numexpr==2.8.1
numpy==1.19.5
oauthlib==3.2.2
packaging==21.3
pandas==1.1.5
pandas-profiling==3.1.0
pandera==0.6.5
pandocfilters==1.5.0
parso==0.7.1
patsy==0.5.3
pexpect==4.8.0
phik==0.12.0
pickleshare==0.7.5
Pillow==8.4.0
plac==1.1.3
platformdirs==2.4.0
plotly==5.15.0
preshed==3.0.8
prometheus-client==0.17.1
prometheus-flask-exporter==0.22.4
prompt-toolkit==3.0.36
protobuf==3.19.6
ptyprocess==0.7.0
pyarrow==1.0.0
pycaret==2.3.10
pycparser==2.21
pydantic==1.9.2
Pygments==2.14.0
PyJWT==2.4.0
pyLDAvis==3.2.2
pynndescent==0.5.10
pyod==1.1.0
pyparsing==3.1.0
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2023.3
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==25.1.0
querystring-parser==1.2.4
ray==2.4.0
regex==2023.6.3
requests==2.27.1
scikit-learn==0.23.2
scikit-plot==0.3.7
scipy==1.5.4
seaborn==0.11.2
Send2Trash==1.8.2
shap==0.41.0
six==1.16.0
slicer==0.0.7
smart-open==6.3.0
smmap==5.0.0
spacy==2.3.9
SQLAlchemy==1.4.49
sqlparse==0.4.4
srsly==1.0.7
statsmodels==0.12.2
tabulate==0.8.10
tangled-up-in-unicode==0.1.0
tenacity==8.2.2
terminado==0.12.1
testpath==0.6.0
textblob==0.17.1
thinc==7.4.6
threadpoolctl==3.1.0
tornado==6.1
tqdm==4.64.1
traitlets==4.3.3
typing_extensions==4.1.1
typing-inspect==0.9.0
umap-learn==0.5.3
urllib3==1.26.16
virtualenv==20.17.1
visions==0.7.4
wasabi==0.10.1
wcwidth==0.2.6
webencodings==0.5.1
websocket-client==1.3.1
Werkzeug==2.0.3
widgetsnbextension==3.6.4
wordcloud==1.9.2
xgboost==1.5.2
yellowbrick==1.3.post1
zipp==3.6.0
